+++ uri/builder.go
@@ -1,4 +1,4 @@
-// Copyright (C) 2025 neocotic
+// Copyright (C) 2026 neocotic
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -25,11 +25,12 @@ package uri
 
 import (
 	"fmt"
-	"github.com/neocotic/go-optional"
 	"maps"
 	"net/url"
 	"slices"
 	"strings"
+
+	"github.com/neocotic/go-optional"
 )
 
 // Builder is used to construct a URI reference.